init SetupSteps components
