+++ src/library/SetupSteps/Footer/Wrapper.ts
@@ -1,0 +1,14 @@
+// Copyright 2022 @paritytech/polkadot-staking-dashboard authors & contributors
+// SPDX-License-Identifier: Apache-2.0
+
+import styled from 'styled-components';
+
+export const Wrapper = styled.div`
+  width: 100%;
+  display: flex;
+  flex-flow: row wrap;
+  align-items: center;
+  justify-content: flex-end;
+  padding: 0 0.25rem;
+  margin-top: 1rem;
+`;
