+++ src/pages/Stake/Wrappers.tsx
@@ -2,7 +2,7 @@
 // SPDX-License-Identifier: Apache-2.0
 
 import styled from 'styled-components';
-import { textSecondary, networkColor } from 'theme';
+import { textSecondary } from 'theme';
 
 export const Wrapper = styled.div`
   display: flex;
@@ -21,62 +21,6 @@ export const StakingAccount = styled.div<{ last: boolean }>`
   h4 {
     color: ${textSecondary};
   }
-`;
-
-export const HeaderWrapper = styled.div`
-  width: 100%;
-  display: flex;
-  flex-flow: row wrap;
-  align-items: center;
-
-  > section {
-    color: ${textSecondary};
-    display: flex;
-    flex-flow: row wrap;
-    align-items: center;
-  }
-
-  > section:last-child {
-    flex: 1;
-    justify-content: flex-end;
-
-    .progress {
-      color: ${textSecondary};
-      opacity: 0.5;
-    }
-
-    .complete {
-      margin: 0;
-      color: ${networkColor};
-    }
-
-    span {
-      margin-right: 1rem;
-    }
-  }
-
-  h2 {
-    margin: 0;
-    padding: 0.3rem 0;
-    display: flex;
-    flex-flow: row wrap;
-    justify-content: flex-start;
-    align-items: center;
-
-    .assistant-icon {
-      margin-left: 0.6rem;
-    }
-  }
-`;
-
-export const FooterWrapper = styled.div`
-  width: 100%;
-  display: flex;
-  flex-flow: row wrap;
-  align-items: center;
-  justify-content: flex-end;
-  padding: 0 0.25rem;
-  margin-top: 1rem;
 `;
 
 export const Section = styled.div`
