+++ src/pages/Stake/types.ts
@@ -1,3 +1,6 @@
+// Copyright 2022 @paritytech/polkadot-staking-dashboard authors & contributors
+// SPDX-License-Identifier: Apache-2.0
+
 export interface GenerateNominationsInnerProps {
   setters: Array<any>;
   nominations: string[];
@@ -8,18 +11,6 @@ export type Nominations = string[];
 
 export interface SetControllerProps {
   section: number;
-}
-
-export interface HeaderProps {
-  title?: string;
-  assistantPage?: string;
-  assistantKey?: string;
-  complete?: boolean | null;
-  thisSection: number;
-}
-
-export interface FooterProps {
-  complete: boolean;
 }
 
 export interface ChooseNominatorsProps {
